# optimized topp/topk calculation